Created SaveData<T> class to contain highScores list
I have implemented the code changes I discussed with ChatGPT.  Specifically the creation of the SaveData<T> class which was able to store a list, which solved the issue of the Json utility not being able to serialize lists (This is because although the json utility can directly serialize a list, it CAN serialize a class which contains a list).  I then made the appropriate modifications to the SaveHighScores() and LoadHighScores() methods so that they were now serializing the new SaveData<HighScore>(highScores) instance of the SaveData class, to a json file (as opposed to attempting to serialize the highScores list directly).  Gosh I feel like I mangled this explanation so feel free to correct any misused terminology or poor phrasing.